[fix] Migrate book API logic to store level for better management
