+++ front-end/src/lib/stores/index.ts
@@ -1,0 +1,8 @@
+import { writable } from 'svelte/store';
+import { libraryStore, libraryAPIStore } from './library';
+import { bookStore } from './book';
+import { notificationStore } from './notification';
+
+export const isLoading = writable<boolean>(true);
+export { libraryStore, libraryAPIStore, bookStore, notificationStore };
+
